[Update] Adição do observer e refatoração da tela de listagem de veículos #19

+++ codigo/business/custos/Custos.java
@@ -2,5 +2,5 @@ package business.custos;
 
 public interface Custos {
 
-    public float calcularCustoTotal();
+    public double calcularCustoTotal();
 }
